--- acslib/base/__init__.py
@@ -1,5 +1,0 @@
-import acslib.base.status as status
-
-from .acs import AccessControlSystem
-from .config import ACSConfig, ACSConfigException
-from .connection import ACSConnection, ACSRequestData, ACSRequestException, ACSRequestResponse
